backend/requirements.txt
absl-py==2.3.1
albucore==0.0.24
albumentations==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
av==15.1.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
celluloid==0.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.3
decorator==4.4.2
deepface==0.0.92
dnspython==2.7.0
easing-functions==1.0.4
easydict==1.13
email_validator==2.2.0
facenet-pytorch==2.6.0
fastapi==0.115.4
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fer==22.5.1
ffmpeg==1.4
filelock==3.19.1
fire==0.7.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.75.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hub-sdk==0.0.24
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
insightface==0.7.3
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.3
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_dtypes==0.5.3
moviepy==1.0.3
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
narwhals==1.48.1
networkx==3.5
nibabel==5.3.2
nilearn==0.12.1
nltools==0.5.1
numexpr==2.8.4
numpy==1.26.4
oauthlib==3.3.1
onnx==1.19.0
onnxruntime==1.22.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.0.74
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==10.2.0
prettytable==3.16.0
proglog==0.1.12
protobuf==6.32.1
psutil==7.0.0
py-cpuinfo==9.0.0
py-feat==0.6.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pynv==0.3
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.12
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
qudida==0.0.4
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.34.1
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
starlette==0.41.3
streamlit==1.47.1
stringzilla==4.0.8
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.9.9
tokenizers==0.21.4
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.55.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.10.0
ultralytics==8.1.0
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.32.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xgboost==3.0.5
